+++ gherkin/common/angle.py
@@ -1,7 +1,4 @@
 from dataclasses import dataclass
-from datetime import datetime
-from enum import Enum
-from typing import NamedTuple, Optional
 
 
 @dataclass(frozen=True, order=True, eq=True)
@@ -36,42 +33,3 @@ class Angle:
         else:
             raise TypeError(f"Expected Angle or in, but got {type(other)}")
         return Angle(new + 360) if new < 0 else Angle(new)
-
-
-class DIRECTION(Enum):
-    CLOCKWISE = "CLOCKWISE"
-    COUNTER_CLOCKWISE = "COUNTER_CLOCKWISE"
-
-
-class SPEED(Enum):
-    FAST = "FAST"
-    FINE = "FINE"
-
-
-@dataclass(order=True, eq=True)
-class Goal:
-    x: int
-    y: int
-    angle: Angle
-
-    def __sub__(self, other):
-        if type(other) == Goal:
-            x_diff = abs(self.x - other.x)
-            y_diff = abs(self.y - other.y)
-            angle_diff = self.angle - other.angle
-            return x_diff, y_diff, angle_diff
-        else:
-            raise TypeError
-
-
-class Rotation(NamedTuple):
-    direction: DIRECTION
-    speed: SPEED
-
-
-class Result(NamedTuple):
-    robot_id: str
-    goal: Goal
-    success: bool
-    completed_at: datetime
-    error: Optional[Exception]
