+++ gherkin/common/enums.py
@@ -1,0 +1,11 @@
+from enum import Enum
+
+
+class DIRECTION(Enum):
+    CLOCKWISE = "CLOCKWISE"
+    COUNTER_CLOCKWISE = "COUNTER_CLOCKWISE"
+
+
+class SPEED(Enum):
+    FAST = "FAST"
+    FINE = "FINE"
